refactored study API routes
