+++ nsweb/api/operations.py
@@ -16,70 +16,6 @@ def make_cache_key():
     return request.path + request.query_string.decode('utf-8')
 
 
-@bp.route('/studies/')
-@cache.cached(timeout=3600, key_prefix=make_cache_key)
-def get_studies():
-    """
-    Retrieve study data
-    ---
-    tags:
-        - studies
-    responses:
-        200:
-            description: Study data
-        default:
-            description: No studies found
-    parameters:
-        - in: query
-          name: limit
-          description: Maximum number of studies to retrieve (default = 25; max = 100)
-          type: integer
-          required: false
-        - in: query
-          name: page
-          description: Page number
-          type: integer
-          required: false
-        - in: query
-          name: search
-          description: Search study fields (title, authors, journal, and year)
-          type: string
-          required: false
-        - in: query
-          name: pmid
-          description: PubMed ID(s) of one or more studies
-          required: false
-          collectionFormat: csv
-          type: array
-          items:
-            type: integer
-    """
-    DEFAULT_LIMIT = 25
-    MAX_LIMIT = 100
-    limit = int(request.args.get('limit', DEFAULT_LIMIT))
-    limit = min(limit, MAX_LIMIT)
-    page = int(request.args.get('page', 1))
-
-    studies = Study.query
-
-    if 'pmid' in request.args or 'id' in request.args:
-        ids = request.args.get('id', None)
-        ids = request.args.get('pmid', ids)
-        ids = re.split('[\s,]+', ids.strip(' ,'))
-        studies = studies.filter(Study.pmid.in_([int(x) for x in ids]))
-
-    if 'search' in request.args and len(request.args['search']) > 1:
-        search = '%{}%'.format(request.args['search'])
-        studies = studies.filter(Study.title.like(search) |
-                                 Study.authors.like(search) |
-                                 Study.journal.like(search) |
-                                 Study.year.like(search))
-
-    studies = studies.paginate(page, limit, False).items
-    schema = StudySchema(many=True)
-    return jsonify(data=schema.dump(studies).data)
-
-
 @bp.route('/decode/')
 @cache.cached(timeout=3600, key_prefix=make_cache_key)
 def get_decoding():
